requirements.txt
fastapi
uvicorn
requests
pydantic
streamlit
python-dotenv
sentence-transformers
chromadb
numpy